requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
firebase-admin==6.8.0
pywebpush==2.0.3
py-vapid==1.9.2
http-ece==1.2.1
paho-mqtt==2.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
pytz==2025.1
tzlocal==5.3.1
APScheduler==3.11.0
pillow==11.1.0
